Support inline sourcemaps without sourcesContent
`lavamoat-browserify` would crash when passed a file with an inline
sourcemap without a `sourcesContent` property. Thus property is
optional, and not guaranteed to be present in a valid sourcemap.

It turns out that the only use of this property was to assign to an
unused variable. This unused variable has been removed, and the plugin
now supports sourcemaps without `sourcesContent`.