chore: replace yarn with npm workspaces
This removes all yarn config, caches, and lockfiles, and changes all scripts to use `npm run` instead of yarn`.  It means `lerna bootstrap` is no longer needed, and `lerna` is not needed (but can still be used) to run scripts across workspaces.  There should only be _one_ lockfile: the root `package-lock.json`.

Note that this change impacts what would be done in #625.  Test fixtures, when installed, should no longer create lockfiles.

A note on "resolutions": these become `overrides` and everything can be done in the monorepo root.

This changeset does not modify CI.