fix: normalize all package versions
Due to the errant release, "newer" versions were published of all of these packages, but VCS did not reflect those new versions.  This updates the versions for each package to the "newest" version and changes the release-please manifest to point to those versions, so that next release will bump them.