chore(github): adds a CODEOWNERS
This adds a `CODEOWNERS` file which assigns reviewers to PRs automatically.  It can also be used for branch protection rules, but we are not doing so currently.

My idea was that if there is a specific area you, as a maintainer, would like to review, then you can add yourself to this file.