feat(aa): add custom resolver
It also accepts a custom resolver, which is needed for actually resolving ESM packages, due to the inability of the `resolve` package to do so.  Adding this natively would break the custom resolver; we may need to abandon it, because `package.json` _must_ be read in order to resolve any package having an `exports` field.