chore(ci): auto-assign PR author
This adds an action to automatically assign (if possible) a PR to the author of that PR. I use this extensively for myself, and it is especially helpful when viewing a list of PRs and filtering them.

The action will not assign a bot to a PR and it will not assign anyone to PRs that already have users assigned.