chore(ci): do not auto-rebase onto main
Renovate was configured to automatically rebase all its PRs onto `main` whenever `main` gets pushed to.  This causes a lot of workflows to get dumped into the CI queue.

This disables that behavior.  I am not sure if it will still automatically attempt to resolve conflicts if things get _too_ out-of-date, but we will see.