chore: add json5 to linter
This adds lint support for JSON5/JSONC files via [eslint-plugin-jsonc](https://npm.im/eslint-plugin-jsonc).

In addition, `tsconfig.json` files are _actually_ treated as JSONC files, regardless of the file extension (this allows us to add comments in them, if we wish; the risk of doing so having negative unintended consequences on 3rd-party tooling is pretty low, given the default `tsconfig.json` generated by `tsc init` contains comments--so it is expected--and _generally_ tools will parse `tsconfig.json` using TS' programmatic API).

The only other affected file at this time is `.github/renovate.json5`. There is currently _no_ validation of this file, and the impact of pushing an invalid config would suck and may not immediately be obvious.