chore(core): run scenarios in ava context
This changes the `runScenarios` test so that there's a 1:1 mapping between each scenario and an Ava test.

This requires changing the file to be `.mjs`, however.

Also: ESLint is mightily confused, and after hammering on it for a little bit, I still couldn't figure out what magic incantations I needed to get it to parse the file.

Take it or leave it!