chore: adapt dirty-check script to be run in local dev env
Hopefully the comments shed some light on what's happening here.

We cannot use `--staged` w/ Git in a local dev env, because it's assumed we're already operating on the stage via `lint-staged`.  Using `--staged` _can and will_ cause us to lose unstaged changes due to the way `lint-staged` works.

The only other thing changed here was how we do logging, since it's weird and wrong to try to use GitHub Action Commands in a dev env.