chore: do not autoformat policy files
Prettier was automatically formatting `node/test/projects/2/atob.policy.json` which keeps causing the file to churn, since when LM writes it, it has no trailing newline. Prettier adds a newline.

This tweaks the `.prettierignore` so that it will not format `*policy.json` instead of just `policy.json`.

It also aligns `.editorconfig` and VS Code to respect this.