fix(core): detect ESM/CJS imports properly
Babel can deduce the source type (CJS/ESM) of the code it's parsed; this lives in the `program.sourceType` nested property of of the `File` object it `parse()` returns. We use this to determine the symbols in the module scope, which should be excluded from the list of globals the inspector finds.