fix(node): better merging of existing overrides with generated policy
This change uses a subclass of `Compartment` during policy generation which forces packages mentioned in a `policy-override.json` into the `imports` array of the `ModuleSource` record within the `ModuleDescriptor` result of `importHook`.  This queues them for linking by `@endo/compartment-mapper` when they would otherwise be ignored.

To work, it requires a change to `captureFromMap()` in `@endo/compartment-mapper` to accept a `Compartment` constructor option, which will then be passed thru to `link()`.

This should be considered experimental 😄