fix(browserify): use @lavamoat/sourcemap-validator
I started investigating an `npm audit` issue and wound up creating a patch for `lodash.template`, but realized that Lego had created a fork of `sourcemap-validator` which removes the `lodash.template` dep. He never got around to actually using it--I am sure that this is what he created it for.