fix(core): ignore Function in GlobalPolicy
This fixes the problems that arise if `globals: {Function: true|'write'}` appears in a `ResourcePolicy`.

Previously, setting this to `true` ("read") would cause the global `Function` wrapping to fail, because the global `Function` would have been replaced with an accessor. If it was set to `write`, it would be wrapped in a _different_ accessor, which would run any script with the incorrect `this` reference.

Policy gen will never create such a policy, which is likely why this bug had been avoided. This means that if we see such a policy, it had to be hand-written with intention (or user error).

With this change, any `globals: {Function: true|'write'}` is ignored.  "Read" access is implied.  It may be prudent to throw instead of ignore the `write` case, as we do not want to allow write permission to the global `Function` anyhow.