fix(node): fix Function constructor behavior in Compartments
This copies some code out of `endowmentsToolkit` (see `function-wrapper.js`) which creates a wrapper for `globalThis.Function` to ensure functions are called with the appropriate context.

We cannot import this function from `endowmentsToolkit` because it runs in a different compartment; the references will not be the same.

Some operations which were previously in `default-attenuator.js` are now in `execution-compartment.js`.  This ensures the operations will be applied to _all_ `Compartment`s; not just the ones with policy.

Renamed `GreedyCompartment` to `PolicyGenCompartment` to differentiate it from `ExecutionCompartment`.