fix(types): Object.hasOwn does not interfere with index signatures
The `Object.hasOwn` type was doing a thing where if you have object `foo` of type `Record<string, any>` and you asked if `Object.hasOwn(foo, 'bar')`, TS would infer that the _value_ of `foo[bar]` is `never` -- which means you cannot subsequently _assign a value_ to `foo[bar]` within a block.

This fixes that problem, but the type is still broken in some circumstances (see note I added to definition).