feat(node): better UX for specifying entrypoints
This allows any "bin script" found in `node_modules` to be specified as the entrypoint _by name_ when coupled with the `--bin`/`-b` flag.

Example:

```sh
lavamoat generate -b grunt
```

This will use `./node_modules/.bin/grunt` as the entrypoint, which will then be resolved to whatever that's symlinked to.

Related changes:

- Move filesystem-related utils into their own module (`fs.js`)
- Create some functions in `fs.js` to aid resolution
- Consume these in `cli.js` which simplifies a few things