feat(core): support generating policies with a "root" field
This adds the `root` field to `LavaMoatPolicy` and updates `lavamoat-core` to support it.  The field is ignored/unused _unless_ the module inspector is created with the `trustRoot = false` option.  If the field is not present, the root will be trusted.  If it is present _and_ has a `usePolicy: some-resource-id`, then it will use that policy instead.

- Adds tests for types, pulling in [expect-type](https://npm.im/expect-type)
- Deprecates `createConfigForTest` and `generateConfigFromFiles` (replace `Config` with `Policy`)
- Removes non-existent snapshots
- Adds `isRoot` field to `LavamoatModuleRecord`.  This field is the source of truth as to whether or not a `LMR` represents the root module.  If its name is `$root$`, this will be true.
- `LavaMoatPolicy` and its ilk can now validate that the `root.usePolicy` value references a key in `resources`