feat(core): support untrusted entrypoints
The policy format such that it now accepts a `root` field which can have an object with a `usePolicy` field. This field references the name (key) of a resource in the `resources` field.

Example:

```json
{
  "resources": {
    "foo": {
      "packages": {
        "bar": true
      }
    }
  },
  "root": {
    "usePolicy": "foo"
  }
}
```

Note that a `usePolicy` field referring to a non-existent resource is an error and will be disallowed by type safety. A `root` field _without_ `usePolicy` is explicitly allowed, but it should be ignored.

Without this field, the root package is considered to have access to everything. The use-case is for executing third-party apps directly using LavaMoat (e.g., build tools).

The internal `LavaMoatModuleRecord` class now has a `isRoot` field which will be set to true if the package name is `$root$` _or_ the instantiating code explicitly sets the flag. Existing implementations do not explicitly set this flag, so new behavior based on the `root.usePolicy` policy field is currently unsupported everywhere except `@lavamoat/node`.

Additionally:

- `resources` is now a required (per types) field of a policy file
- the previous item makes `LavaMoatOverrides` and `PartialLavaMoatPolicy` useless as overrides now do not differ from policy; removed
- `LavaMoatPolicy` accepts a type argument `T` which is a `Resources` type. This enables validation of the `root.usePolicy` value, since it must reference a property in `Resources`.
- fix some internal type problems and shortcomings
- deprecate internal test utils referencing "config" in lieu of those referencing "policy"
- `expect-type` used for asserting expected behavior of types
- removed unused `type-fest` from prod dependencies