feat(node): support untrusted entrypoints
This adds support for "untrusted entrypoints"; execution of third-party code against a limited policy. The behavior of legacy `lavamoat-node` is to run the root package where everything is allowed. This is (arguably) appropriate for first-party code, but for _third_-party code it can be an unacceptable risk.

`@lavamoat/node` now detects if the _real path_ of the entrypoint lives in a `node_modules` folder. If it does, it is assumed to be third-party code, and will be treated as such. This means that during policy generation, the package of the entrypoint will be treated just like any other package. At runtime, `@lavamoat/node` respects the policy as defined.

To define a policy with an untrusted root package, the `root.usePolicy` field must be present in the policy (or policy override) file. It must refer to a key in the `resources` field of policy; this policy will be used for the root package.

If a policy has a valid `root.usePolicy` field and the entrypoint refers to a package which is _not_ in `node_modules`, then `@lavamoat/node` will exit with an error stating that there's a root trust mismatch. Likewise, if a policy lacks a valid `root.usePolicy` field and the entrypoint _is_ in a `node_modules` folder, it will exit with a similar error.  This provides some guarantee that execution is as safe _as expected._

Additionally, there were a handful of bugfixes applied:

- `--dev` flag now fully respected
- support for packages providing both conditional subpath exports with the `node` condition _and_ a top-level `browser` field
- policy override expansion works appropriately
- policy generation won't notify the user a policy has been written when it has not
- upgrades to `@endo/evasive-transform` and `@endo/compartment-mapper` address some ecosystem incompatibilities
- support for JS sources starting with hashbangs

And:

- better SES incompatibility warning output
- some updates to public API allowing for providing powers ("capabilities") and either raw policy/overrides _xor_ a path/URL to policy/overrides