chore: update CODEOWNERS
This removes `@LavaMoat/devs` from the `CODEOWNERS` file.  The problem is that if _anyone_ reviews a given PR, then the review request is considered "completed"; the other team members will no longer have an open review request.

Instead, we can add individual team members directly to all PRs. This eliminates the need for path-specific ownership (which we are too small to really need).