chore(ci): change how project review workflow works
This changes the project review workflow so that it _only triggers once, at time of PR creation_.

It does not trigger for Renovate-created PRs.

To support this, the PR review board has been modified to exclude "draft" PRs.

* * *

Ideally, we want to ensure that draft PRs do not clog up the review board. The reason for this change is mostly that _we cannot reliably keep a PR "in sync" with the review board_ depending on its draft status.  Further, attempting to add the same PR multiple times will result in a worfklow error.  This sidesteps the issue by adding _all_ non-bot PRs to the review board, where we can easily filter out draft PRs.