chore(ci): remove pr-review workflow
GH Projects can be configured to automatically add PRs to themselves when certain conditions are met.

I have configured the PR review board to automatically add any _non-draft PRs having an assignee_ to itself. PRs from Renovate do not satisfy this filter, since they are skipped by the [toshimaru/auto-author-assign](https://github.com/toshimaru/auto-author-assign) action.  Any new PR _from a maintainer_ will automatically be assigned to their PR.  Unfortunately, Projects have a limited set of fields they can query (e.g., they cannot query the PR _author_), so this is what I came up with for a workaround.

If we remove this workflow and the automation is unsuccessful, we can restore it—but it seems to me that is both safer and faster to avoid GHA if we can.