chore(aa): cleanup symlinks
This was found when running CI locally (self-hosted). A existing symlink created outside of the self-hosted container causes the test to fail. This makes tests using `createProject4Symlink` cleanup after themselves _and_ ensures a clean symlink upon each execution.

We _also_ need comprehensive "clean" scripts, but that's beyond the scope of this change. 😄