chore(node): avoid mutating CompartmentDescriptor.label
Given the realization that mutating `CompartmentDescriptor.label` is unwise (as `@endo/compartment-mapper`) depends upon it), we need another way to associate our own metadata (e.g., the canonical name) with a compartment descriptor. This change decouples this metadata by storing it in a separate data structure which can then be referenced based on the compartment descriptor's `location` property (which is also equal to the "compartment descriptor name"; the key of the `CompartmentMapDescriptor.compartments` record—_not_ the `CompartmentDescriptor.name` field).

Conceptually, we destroy the recently-introduced `CompartmentDescriptorTransform` concept (which allowed modular, direct mutation of compartment map descriptors) and replace it with a `CompartmentDescriptorDecorator` (name subject to change). This works like a transformer _on metadata_—_not_ the compartment descriptor itself.  For each compartment descriptor, a "decorator" receives a (possibly incomplete) metadata object and returns a new object (or the same one, I guess) containing whatever it needed to add to the metadata. There's only one so far, which stores the canonical name. Programmatic consumers may add their own metadata using this architecture.

We can then provide this mapping of metadata to operations that need it (mainly during policy generation).

This change also reduces some boilerplate around calling `mapNodeModules()`, since both execution and policy generation call it nearly identically. The new function is `makeNodeCompartmentMap()`.

`makeNodeCompartmentMap()` will generate the initial compartment descriptor metadata and assert that it is valid. For policy generation, we call `captureFromMap()` to "normalize" or "distill" the `CompartmentMapDescriptor` such that it no longer references filepaths and can be used with policy.  This necessitates a remapping of the metadata (and subsequent validation) since the compartment descriptor `location` field has changed.  Note that we _must_ initially compute the canonical names post-dependency-graphing, since the `path` field is stripped from each `CompartmentDescriptor` by `captureFromMap()` and it is no longer possible to compute.