chore(ci): add a thing that updates coverage thresholds
For workspaces which have a `test:cover` script--presumably which uses `c8`--this automatically updates the `.c8rc` in each workspace if the coverage check passes. `.c8rc` controls the coverage threshold, which must pass on subsequent executions of `c8 --check-coverage`.

It is build as a scheduled or on-demand workflow which only works on the main branch.
I am not super enthusiastic about having commits automatically pushed to `main`. This is just one way to do it.

An alternative would be to automatically update `.c8rc` with the relevant contents of `coverage/coverage-summary.json` as a pre-commit hook; this is maybe a better idea, but we would likely want to write a custom JS script in lieu of `.github/actions/coverage-threshold/action.yml`.

The point is this: _if_ we want to maintain a high level of code coverage, we do not want it to ever drop. Developers will not remember to manually update the thresholds, so it should be automated. This is an experiment.