chore(node): refactor CLI
The CLI was getting unwieldy so I split out the middleware functions and options.

Most options which were previously global are no longer, because they are not guaranteed to be used across all commands.  The only ones which remain are `verbose` and `quiet`, which are now in the "logging" option group.