fix(node): fix import map resolution
When building import maps for `LavamoatModuleRecord`s, we were wildly guessing at the actual filepath for any given module descriptor. This change employs a function to use the information from the `CompartmentMapDescriptor` (coupled with the rename mapping, which we need to do just about anything with paths) to query the appropriate compartment for the module descriptor which matches its specifier, finally building a path based on the real compartment descriptor path and module.

# Conflicts:
#	packages/node/src/policy-gen/policy-gen-context.js