fix(node): correct policy loading problems
There was a bug where providing an explicit policy path would make a poor guess about the policy override path.  Now, the default policy override path is computed as a sibling file to the policy path. If the policy path is not handy, then it is relative to the project root and `lavamoat/node`. The project root defaults to cwd.