chore(ci): update release-please config with package names (#1727)
This adds the (oddly missing-from-schema) `package-name` field to each "component" in `release-please-config.json`. I believe this will both change the release title _and_ the tag names going forward.

As such, we may have a concern that legacy `lavamoat` (`lavamoat-node`) will now be named `lavamoat`.  Maybe we should change this to be `lavamoat-node` or `lavamoat-legacy` or smth.