chore(policy): add a moatload of integration tests
This copies all extant `policy.json` files from other packages' examples and test fixtures and validates them using the schemas.

The only one that is expected to fail is `policy-3.json`, which contains a resource by the name of `<root>`, which is invalid.  The proper way to express this would be to create a resource with the (canonical) name of the root package, and then use the `root.usePolicy` property to reference the name.